frequency: new utils.py file
-functions not related to the bottle are moved to a separate file utils.py
-formatted ny black
